refactor(weight): modularize weight evolution logic, extract chart/view components, and fix progress edge case
